src/requirements.txt
bleach==3.1.0
Django==2.1.7
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
flake8==3.5.0
flake8-isort==2.6.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
isort==4.3.4
psycopg2-binary==2.8.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
pytest-flake8==1.0.2
pytest-forked==0.2
pytest-mock==1.10.0
pytest-xdist==1.24.1
setuptools==40.6.2
